refactor: consoleInput 분리
