+++ src/main/java/racingcar/RacingCarGame.java
@@ -2,7 +2,7 @@ package racingcar;
 
 public class RacingCarGame {
     public static void main(String[] args) {
-        UserInput userInput = UserInput.getUserInputFromScanner();
+        UserInput userInput = ConsoleInput.getUserInput();
         Race race = userInput.readyForRace();
         race.start();
     }
